+++ docs/auth0_users_blocks.md
@@ -8,4 +8,5 @@ Manage brute-force protection user blocks.
 ## Commands
 
 - [auth0 users blocks list](auth0_users_blocks_list.md) - List brute-force protection blocks for a given user
+- [auth0 users blocks unblock](auth0_users_blocks_unblock.md) - Remove brute-force protection blocks for a given user
