+++ docs/auth0_users_delete.md
@@ -51,7 +51,6 @@ auth0 users delete [flags]
 - [auth0 users roles](auth0_users_roles.md) - Manage a user's roles
 - [auth0 users search](auth0_users_search.md) - Search for users
 - [auth0 users show](auth0_users_show.md) - Show an existing user
-- [auth0 users unblock](auth0_users_unblock.md) - Remove brute-force protection blocks for a given user
 - [auth0 users update](auth0_users_update.md) - Update a user
